update tests and yaml
