+++ tests/unit/backend/cellguide/api/test_marker_genes.py
@@ -1,19 +1,7 @@
 from unittest.mock import patch
 
 from tests.unit.backend.layers.common.base_api_test import BaseAPIPortalTest
-
-
-def mock_get_marker_gene_data():
-    return {
-        "Homo sapiens": {
-            "brain": {
-                "CL:0000540": [
-                    {"marker_score": 0.95, "me": 0.5, "pc": 0.1, "gene": "Gene1"},
-                    {"marker_score": 0.90, "me": 0.4, "pc": 0.2, "gene": "Gene2"},
-                ]
-            }
-        }
-    }
+from tests.test_utils.mocks import mock_get_marker_gene_data
 
 
 class TestMarkerGenesAPI(BaseAPIPortalTest):
@@ -26,11 +14,31 @@ class TestMarkerGenesAPI(BaseAPIPortalTest):
         new=mock_get_marker_gene_data,
     )
     def test_get_marker_genes(self):
+        # Test case: Organism, tissue, and cell type provided
         response = self.app.get(
             "/cellguide/v1/marker_genes?organism=NCBITaxon:9606&tissue=UBERON:0000955&cell_type=CL:0000540"
         )
         self.assertEqual(response.status_code, 200)
         data = response.get_json()
-        self.assertEqual(len(data), 2)
-        self.assertEqual(data[0]["gene"], "Gene1")
-        self.assertEqual(data[1]["gene"], "Gene2")
+        self.assertEqual(data, self.mock_marker_gene_data["Homo sapiens"]["brain"]["CL:0000540"])
+
+        # Test case: Only organism is provided
+        response = self.app.get("/cellguide/v1/marker_genes?organism=NCBITaxon:9606")
+        self.assertEqual(response.status_code, 200)
+        data = response.get_json()
+        expected_data = self.mock_marker_gene_data["Homo sapiens"]
+        self.assertEqual(data, expected_data)
+
+        # Test case: Organism and tissue is provided
+        response = self.app.get("/cellguide/v1/marker_genes?organism=NCBITaxon:9606&tissue=brain")
+        self.assertEqual(response.status_code, 200)
+        data = response.get_json()
+        expected_data = self.mock_marker_gene_data["Homo sapiens"]["brain"]
+        self.assertEqual(data, expected_data)
+
+        # Test case: Organism and cell type is provided, agnostic to tissues
+        response = self.app.get("/cellguide/v1/marker_genes?organism=NCBITaxon:9606&cell_type=CL:0000540")
+        self.assertEqual(response.status_code, 200)
+        data = response.get_json()
+        expected_data = self.mock_marker_gene_data["Homo sapiens"]["All Tissues"]["CL:0000540"]
+        self.assertEqual(data, expected_data)
